% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rut.R
\name{plotter}
\alias{plotter}
\title{Get the Rut work history and prepare for plotting}
\usage{
plotter(database, workhistory)
}
\arguments{
\item{database}{This is the output of the treat_assign function}

\item{workhistory}{This is the output of the processing function}
}
\description{
Get the Rut work history and prepare for plotting
}
