% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rut.R
\name{pmis_preprocessing}
\alias{pmis_preprocessing}
\title{Get the preprocessing done}
\usage{
pmis_preprocessing(database, corrected_DFOS, min_year, max_year, asphalt = T)
}
\arguments{
\item{database}{This is the full PMIS work history stored in the UT Database HUB}

\item{corrected_DFOS}{This is the database with the corrected DFOs found in the UT Database Hub}

\item{min_year}{This is the minimum year you want to analyze, we recommend 2018}

\item{max_year}{This is the maximum year you want to analyze, we recommend 2019}

\item{asphalt}{This parameter double checks that you are analyzing HMAs, by default it is true}
}
\description{
Get the preprocessing done
}
