% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rut.R
\name{pmis_processing}
\alias{pmis_processing}
\title{Get the Rut work history}
\usage{
pmis_processing(database)
}
\arguments{
\item{database}{This is the output of the preprocessing function}
}
\description{
Get the Rut work history
}
