% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rut.R
\name{treat_assing}
\alias{treat_assing}
\title{Get the Rut work history}
\usage{
treat_assing(database, workhistory)
}
\arguments{
\item{database}{This is the output of the preprocessing function}

\item{workhistory}{This is the output of the processing function}
}
\description{
Get the Rut work history
}
